\name{ATA-package}
\alias{ATA-package}
\alias{ATA}
\docType{package}
\title{
\packageTitle{ATA}
}
\description{
\packageDescription{ATA}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ATA}
\packageIndices{ATA}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ATA}

Maintainer: \packageMaintainer{ATA}
}
\references{
~~ Literature or other references for background information ~~
}
%%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
