% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeconstobj.R
\name{makeconstobj}
\alias{makeconstobj}
\title{User-guided Function to Create a Constraints Input for ATA Form Creation}
\usage{
makeconstobj( ipool,
 id,
 empty = FALSE)
}
\arguments{
\item{ipool}{Item by characteristic (or property) metadata.}

\item{id}{Name (not actual codes) of unique item identifier (variable).}

\item{empty}{Should the function return an empty list to be filled in manually. Default is \code{FALSE}.}
}
\value{
A list object with "nC" 
                           "nameC" 
                           "lowerC" 
                           "upperC" 
                           "wC"
                           "nI"
                           "set_id"
}
\description{
Guides the user to create a complex list object identifying the constraints to be applied in automated test assembly functions from the \code{ata} package.
}
\references{
Parshall, C. G., et al. (2002). Automated test assembly for online administration. In C. G. Parshall, J. A. Spray, J. C. Kalohn, & T. Davey, Practical considerations in computer based testing (pp.106-125). New York, NY: Springer-Verlag New York, Inc.
}
\author{
Gulsah Gurkan (gurkangulsah@gmail.com), Michael Chajewski (mchajewski@hotmail.com)
}
\keyword{ata}
\keyword{automatest_test_assembly}
\keyword{lp}
\keyword{test_form}
\keyword{test_via_lp}
\keyword{test_via_wdm}
\keyword{testform}
\keyword{wdm}
