% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_large_example.R
\docType{data}
\name{metadata_large_example}
\alias{metadata_large_example}
\title{ATA Package Large Example Item Metadata}
\format{A data frame with 1096 rows and 44 variables:
\describe{
  \item{Item}{Unique item identifier, alpha-numeric index code.}
  \item{Content}{Content label, as factor identifying content "A" and "B".}
  \item{Content_A}{Dummy code for content "A", 0 and 1 indicators.}
  \item{Content_B}{Dummy code for content "B", 0 and 1 indicators.}
  \item{Content_C}{Dummy code for content "C", 0 and 1 indicators.}
  \item{Content_D}{Dummy code for content "D", 0 and 1 indicators.}
  \item{Content_E}{Dummy code for content "E", 0 and 1 indicators.}
  \item{Content_F}{Dummy code for content "F", 0 and 1 indicators.}
  \item{Content_G}{Dummy code for content "G", 0 and 1 indicators.}
  \item{Content_H}{Dummy code for content "H", 0 and 1 indicators.}
  \item{Content_I}{Dummy code for content "I", 0 and 1 indicators.}
  \item{Content_J}{Dummy code for content "J", 0 and 1 indicators.}
  \item{Content_K}{Dummy code for content "K", 0 and 1 indicators.}
  \item{Content_L}{Dummy code for content "L", 0 and 1 indicators.}
  \item{Content_M}{Dummy code for content "M", 0 and 1 indicators.}
  \item{Content_N}{Dummy code for content "N", 0 and 1 indicators.}
  \item{Content_O}{Dummy code for content "O", 0 and 1 indicators.}
  \item{Content_P}{Dummy code for content "P", 0 and 1 indicators.}
  \item{Content_Q}{Dummy code for content "Q", 0 and 1 indicators.}
  \item{Content_R}{Dummy code for content "R", 0 and 1 indicators.}
  \item{Content_S}{Dummy code for content "S", 0 and 1 indicators.}
  \item{Content_T}{Dummy code for content "T", 0 and 1 indicators.}
  \item{Content_U}{Dummy code for content "U", 0 and 1 indicators.}
  \item{Content_V}{Dummy code for content "V", 0 and 1 indicators.}
  \item{Content_W}{Dummy code for content "W", 0 and 1 indicators.}
  \item{Content_X}{Dummy code for content "X", 0 and 1 indicators.}
  \item{Content_Y}{Dummy code for content "Y", 0 and 1 indicators.}
  \item{Content_Z}{Dummy code for content "Z", 0 and 1 indicators.}
  \item{p}{Item proportion correct responding.}
  \item{rpbis}{Item-total point biserial correlation, rounded decimal correlation in range -1.00 to 1.00}
  \item{iSx}{Item contribution to total composite standard deviation.}
  \item{Time}{Observed median item response time, in seconds.}
  \item{Choices}{Number of response choices in the multiple-choice question.}
  \item{Answer}{Correct answer key. In the multiple-choice questions answer 1 = A, 2 = B, and so on.}
  \item{Skill}{Formative insight skill classification.}
  \item{Skill_1}{Formative skill insight dummy code for skill 1 (S1): Interpretive.}
  \item{Skill_2}{Formative skill insight dummy code for skill 2 (S2): Factual.}
  \item{Skill_3}{Formative skill insight dummy code for skill 3 (S3): Evaluative.}
  \item{IIF_m2}{Item Response Theory (IRT) item information function value at a theta = -2.0.}
  \item{IIF_m1}{Item Response Theory (IRT) item information function value at a theta = -1.0.}
  \item{IIF_0}{Item Response Theory (IRT) item information function value at a theta = 0.0.}
  \item{IIF_1}{Item Response Theory (IRT) item information function value at a theta = 1.0.}
  \item{IIF_2}{Item Response Theory (IRT) item information function value at a theta = 2.0.}
  \item{Parent0}{Passage based item set parent identification.}
}}
\usage{
metadata_large_example
}
\description{
Sample data used to demonstrate automated test assembly.
}
\keyword{datasets}
